pyproject.toml
[project]
name = "temporal-videogen"
version = "0.1.0"
description = "Generate videos using Temporal, Google Gemini, and Veo 2."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage==3.1.1",
    "google-genai[aiohttp]==1.21.0",
    "moviepy==2.2.1",
    "pydantic==2.11.7",
    "structlog==25.4.0",
    "temporalio==1.12.0",
]
